Feature ETP-83: Fix TS errors
