+++ packages/MainUI/package.json
@@ -6,7 +6,7 @@
   "proxy": "http://localhost:8092",
   "scripts": {
     "dev": "vite",
-    "build": "vite build",
+    "build": "tsc && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview"
   },
