+++ packages/storybook/src/stories/Components/Table/types.ts
@@ -26,7 +26,6 @@ export interface TableProps {
   data: Organization[];
   isTreeStructure?: boolean;
   customLabels?: Record<string, string>;
-  isDrawerOpen: boolean;
 }
 
 export interface SidebarContentProps {
